+++ src/imio/esign/config.py
@@ -39,6 +39,10 @@ def get_esign_registry_max_session_size(default=100):
     return api.portal.get_registry_record("imio.esign.max_session_size", default=default)
 
 
+def get_esign_registry_max_session_files(default=25):
+    return api.portal.get_registry_record("imio.esign.max_session_files", default=default)
+
+
 def get_esign_registry_external_watchers():
     value = api.portal.get_registry_record("imio.esign.external_watchers", default="")
     if not value:
@@ -80,6 +84,10 @@ def set_esign_registry_signing_users_email_content(value):
 
 def set_esign_registry_max_session_size(value):
     api.portal.set_registry_record("imio.esign.max_session_size", value)
+
+
+def set_esign_registry_max_session_files(value):
+    api.portal.set_registry_record("imio.esign.max_session_files", value)
 
 
 def set_esign_registry_external_watchers(value):